Renaming in some classes
Renaming in some classes and packages for better readability
