+++ java/src/lab5/task_5_1/Main.java
@@ -1,15 +1,15 @@
 package lab5.task_5_1;
 
-import lab5.task_5_1.freader.FileDataReader;
+import lab5.task_5_1.fs.FileSystem;
 import lab5.task_5_1.strprocessor.*;
 
 public class Main {
 
 	public static void main(String[] args) {
-		FileDataReader fr = new FileDataReader();
+		FileSystem fs = new FileSystem();
 		String data = null;
 		try {
-			data = fr.read("./src/lab5/task_5_1/test_files/test.txt");
+			data = fs.read("./src/lab5/task_5_1/test_files/test.txt");
 		} catch(Exception ex) {
 			System.out.print(1);
 			System.out.println(ex.getMessage());
